Fix #9 (Item Serialization Issues)
tldr; When loading ECHO Themes, the inventory key will be `null` because I dont see any inventory key that they fall into. Hindsight tells me that if they don't have any parts, it doesn't explicitly matter and so that should be fine... But I also don't quite feel like implementing that yet for this issue. (Plus this fix is correct anyway)

So now I backup the original data when loading from a profile (using a protobuf struct from the Character object), and will save that out when unable to serialize an item properly.